Successfully developed course builder form

+++ server/controllers/Section.js
@@ -1,5 +1,6 @@
 const Section = require("../models/Section");
 const Course = require("../models/Course");
+const SubSection = require("../models/SubSection");
 // CREATE a new section
 exports.createSection = async (req, res) => {
 	try {
@@ -54,15 +55,24 @@ exports.createSection = async (req, res) => {
 // UPDATE a section
 exports.updateSection = async (req, res) => {
 	try {
-		const { sectionName, sectionId } = req.body;
+		const { sectionName, sectionId, courseId } = req.body;
 		const section = await Section.findByIdAndUpdate(
 			sectionId,
 			{ sectionName },
 			{ new: true }
 		);
+
+		const course = await Course.findById(courseId).populate({
+			path : "courseContent",
+			populate : {
+				path: "subSection",
+			},
+		}).exec();
+
 		res.status(200).json({
 			success: true,
-			message: section,
+			message: "section updated successfully",
+			data: course,
 		});
 	} catch (error) {
 		console.error("Error updating section:", error);
@@ -77,12 +87,39 @@ exports.updateSection = async (req, res) => {
 exports.deleteSection = async (req, res) => {
 	try {
 		//HW -> req.params -> test
-		const { sectionId } = req.params;
+		const { sectionId, courseId } = req.body;
+		await Course.findByIdAndUpdate(courseId,{
+			$pull : {
+				courseContent : sectionId,
+			}
+		})
+
+		const section = await Section.findById(sectionId);
+		if(!section){
+			return  res.status(404).json({
+				success:false,
+				message: 'No such section exists'
+			})
+		}
+
+		// Find all subsections of the section
+		const subsectionsToDelete = section.subSection;
+		//delete sub-section
+		await SubSection.deleteMany({ _id: { $in: subsectionsToDelete } });
 		await Section.findByIdAndDelete(sectionId);
-		//HW -> Course ko bhi update karo
+
+		//find course and update
+		const  updatedCourse = await Course.findById(courseId).populate({
+			path: "courseContent",
+			populate: {
+				path : "subSection"
+			}
+		}).exec();
+
 		res.status(200).json({
 			success: true,
-			message: "Section deleted",
+			message: "Section deleted successfully",
+			data : updatedCourse,
 		});
 	} catch (error) {
 		console.error("Error deleting section:", error);
